Internal Documentation

Documentation for EpiInfModels.jl's internal interface.

Contents

Contents

Index

Internal API

Index

EpiAware.EpiAwareBase.generate_latent_infs

EpiAware.EpiAwareBase.generate_latent_infs

EpiAware.EpiAwareBase.generate_latent_infs

EpiAware.EpiInfModels.make_renewal_init

EpiAware.EpiInfModels.make_renewal_init

EpiAware.EpiInfModels.neg_MGF

Internal API

EpiAware.EpiAwareBase.generate_latent_infs — Method

generate_latent_infs(
    epi_model::AbstractTuringRenewal,
    _Rt
) -> Any

Implement the generate_latent_infs function for the Renewal model.

Example usage with Renewal type of model for unobserved infection process

generate_latent_infs can be used to construct a Turing model for the latent infections conditional on the sample path of a latent process. In this example, we generate a sample of a white noise latent process.

First, we construct an Renewal struct with an EpiData object, an initialisation prior and a transformation function.

using Distributions, Turing, EpiAware
gen_int = [0.2, 0.3, 0.5]
g = exp

# Create an EpiData object
data = EpiData(gen_int, g)

# Create an Renewal model
renewal_model = Renewal(data = data, initialisation_prior = Normal())

Then, we can use generate_latent_infs to construct a Turing model for the unobserved infection generation model set by the type of renewal_model.

# Construct a Turing model
Z_t = randn(100) * 0.05
latent_inf = generate_latent_infs(renewal_model, Z_t)

Now we can use the Turing PPL API to sample underlying parameters and generate the unobserved infections.

# Sample from the unobserved infections model

#Sample random parameters from prior
θ = rand(latent_inf)
#Get unobserved infections as a generated quantities from the model
I_t = generated_quantities(latent_inf, θ)

EpiAware.EpiAwareBase.generate_latent_infs — Method

generate_latent_infs(
    epi_model::DirectInfections,
    Z_t
) -> Any

Implement the generate_latent_infs function for the DirectInfections model.

Example usage with DirectInfections type of model for unobserved infection process

First, we construct a DirectInfections struct with an EpiData object, an initialisation prior and a transformation function.

using Distributions, Turing, EpiAware
gen_int = [0.2, 0.3, 0.5]
g = exp

# Create an EpiData object
data = EpiData(gen_int, g)

# Create a DirectInfections model
direct_inf_model = DirectInfections(data = data, initialisation_prior = Normal())

Then, we can use generate_latent_infs to construct a Turing model for the unobserved infection generation model set by the type of direct_inf_model.

# Construct a Turing model
Z_t = randn(100)
latent_inf = generate_latent_infs(direct_inf_model, Z_t)

Now we can use the Turing PPL API to sample underlying parameters and generate the unobserved infections.

# Sample from the unobserved infections model

#Sample random parameters from prior
θ = rand(latent_inf)
#Get unobserved infections as a generated quantities from the model
I_t = generated_quantities(latent_inf, θ)

EpiAware.EpiAwareBase.generate_latent_infs — Method

generate_latent_infs(epi_model::ExpGrowthRate, rt) -> Any

Implement the generate_latent_infs function for the ExpGrowthRate model.

Example usage with ExpGrowthRate type of model for unobserved infection process

generate_latent_infs can be used to construct a Turing model for the latent infections conditional on the sample path of a latent process. In this example, we generate a sample of a white noise latent process.

First, we construct an ExpGrowthRate struct with an EpiData object, an initialisation prior and a transformation function.

using Distributions, Turing, EpiAware
gen_int = [0.2, 0.3, 0.5]
g = exp

# Create an EpiData object
data = EpiData(gen_int, g)

# Create an ExpGrowthRate model
exp_growth_model = ExpGrowthRate(data = data, initialisation_prior = Normal())

Then, we can use generate_latent_infs to construct a Turing model for the unobserved infection generation model set by the type of direct_inf_model.

# Construct a Turing model
Z_t = randn(100) * 0.05
latent_inf = generate_latent_infs(exp_growth_model, Z_t)

Now we can use the Turing PPL API to sample underlying parameters and generate the unobserved infections.

# Sample from the unobserved infections model

#Sample random parameters from prior
θ = rand(latent_inf)
#Get unobserved infections as a generated quantities from the model
I_t = generated_quantities(latent_inf, θ)

EpiAware.EpiInfModels.make_renewal_init — Method

make_renewal_init(epi_model::Renewal, I₀, Rt) -> Any

Create the initial vector of infected individuals for a renewal model.

Arguments

epi_model::Renewal: The renewal model.

I₀: The initial number of infected individuals.

Rt: The time-varying reproduction number.

Returns

The initial vector of infected individuals.

EpiAware.EpiInfModels.make_renewal_init — Method

make_renewal_init(
    epi_model::RenewalWithPopulation,
    I₀,
    Rt
) -> Tuple{Any, Float64}

Constructs the initial conditions for a renewal model with population.

Arguments

epi_model::RenewalWithPopulation: The renewal model with population.

I₀: The initial number of infected individuals.

Rt: The time-varying reproduction number.

Returns

A tuple containing the initial number of infected individuals at each generation

interval and the population size of susceptible/available people.

EpiAware.EpiInfModels.neg_MGF — Method

neg_MGF(r, w::AbstractVector) -> Any

Compute the negative moment generating function (MGF) for a given rate r and weights w.

Arguments

r: The rate parameter.

w: An abstract vector of weights.

Returns

The value of the negative MGF.