internal Documentation

Documentation for EpiAware.jl's internal interface.

Contents

Index

EpiAware.NegativeBinomialMeanClust

EpiAware.generate_observation_kernel

EpiAware.neg_MGF

Internal API

EpiAware.NegativeBinomialMeanClust — Method

NegativeBinomialMeanClust NegativeBinomialMeanClust(μ, α)

Compute the mean-cluster factor negative binomial distribution.

Arguments

μ: The mean of the distribution.

α: The clustering factor parameter.

Returns

A NegativeBinomial distribution object.

Signatures

NegativeBinomialMeanClust(
    μ,
    α
) -> Distributions.NegativeBinomial

Methods

NegativeBinomialMeanClust(μ, α)

defined at /home/runner/work/Rt-without-renewal/Rt-without-renewal/src/utilities.jl:210.

EpiAware.generate_observation_kernel — Method

generateobservationkernel generateobservationkernel(delayint, timehorizon)

Generate an observation kernel matrix based on the given delay interval and time horizon.

Arguments

delay_int::Vector{Float64}: The delay PMF vector.

time_horizon::Int: The number of time steps of the observation period.

Returns

K::SparseMatrixCSC{Float64, Int}: The observation kernel matrix.

Signatures

generate_observation_kernel(delay_int, time_horizon) -> Any

Methods

generate_observation_kernel(delay_int, time_horizon)

defined at /home/runner/work/Rt-without-renewal/Rt-without-renewal/src/utilities.jl:229.

EpiAware.neg_MGF — Method

negMGF negMGF(r, w::AbstractVector)

Compute the negative moment generating function (MGF) for a given rate r and weights w.

Arguments

r: The rate parameter.

w: An abstract vector of weights.

Returns

The value of the negative MGF.

Signatures

neg_MGF(r, w::AbstractVector) -> Any

Methods

neg_MGF(r, w)

defined at /home/runner/work/Rt-without-renewal/Rt-without-renewal/src/utilities.jl:132.