Public Documentation

Documentation for EpiAware.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

Contents

Contents

Index

Public API

Index

EpiAware.EpiAware

EpiAware.AbstractEpiModel

EpiAware.DirectInfections

EpiAware.EpiData

EpiAware.ExpGrowthRate

EpiAware.Renewal

EpiAware.Renewal

EpiAware.R_to_r

EpiAware.create_discrete_pmf

EpiAware.create_discrete_pmf

EpiAware.generate_latent_infs

EpiAware.generate_latent_infs

EpiAware.generate_latent_infs

EpiAware.generate_latent_infs

EpiAware.manypathfinder

EpiAware.r_to_R

EpiAware.scan

EpiAware.spread_draws

Public API

EpiAware.EpiAware — Module

module EpiAware

EpiAware provides functionality for fitting epidemiological models to data. It is built on top of the Turing probabilistic programming language, and provides a set of utilities for constructing and fitting models to data.

Core model structure

An epidemiological model in EpiAware consists of composable structs with core abstract types. The core types are:

AbstractModel: This overarching type is used to abstract Turing models and is inheried by the other abstract types we use.

AbstractEpiModel: Subtypes of this abstract type represent different models for the spread of an infectious disease. Each model type has a corresponding make_epi_aware function that constructs a Turing model for fitting the model to data. Implemented concrete subtypes:

Renewal: A renewal process model for the spread of an infectious disease.

ExpGrowthRate: An exponential growth rate model for the spread of an infectious disease.

DirectInfections: A model for the spread of an infectious disease based on modelling direct infections.

AbstractLatentModel: Subtypes of this abstract type represent different latent processes that can be used in an epidemiological model. Implemented concrete subtype:

RandomWalk: A random walk latent process.

AbstractObservationModel: Subtypes of this abstract type represent different observation models that can be used in an epidemiological model. Implemented concrete subtypes:

DelayObservation: An observation process that models the delay between the time of infection and the time of observation as a convolution, followed by a negative binomial distributed sample.

EpiAware.AbstractEpiModel — Type

abstract type AbstractEpiModel <: AbstractModel

abstract type AbstractEpiModel <: AbstractModel end

The abstract supertype for all structs that define a model for generating unobserved/latent infections.

Fields

EpiAware.DirectInfections — Type

struct DirectInfections{S<:Distributions.Sampleable} <: AbstractEpiModel

Model unobserved/latent infections as a transformation on a sampled latent process.

Mathematical specification

If $Z_t$ is a realisation of the latent model, then the unobserved/latent infections are given by

\[I_t = g(\hat{I}_0 + Z_t).\]

where $g$ is a transformation function and the unconstrained initial infections $\hat{I}_0$ are sampled from a prior distribution, initialisation_prior which must be supplied to the DirectInfections constructor. The default initialisation_prior is the standard Normal Distributions.Normal().

Constructor

DirectInfections can be constructed by passing an EpiData object and subtype of Distributions.Sampleable.

Example usage with generate_latent_infs

generate_latent_infs can be used to construct a Turing model for the latent infections conditional on the sample path of a latent process. In this example, we generate a sample of a white noise latent process.

First, we construct a DirectInfections struct with an EpiData object, an initialisation prior and a transformation function.

using Distributions, Turing, EpiAware
gen_int = [0.2, 0.3, 0.5]
g = exp

# Create an EpiData object
data = EpiData(gen_int, g)

# Create a DirectInfections model
direct_inf_model = DirectInfections(data = data, initialisation_prior = Normal())

Then, we can use generate_latent_infs to construct a Turing model for the unobserved infection generation model set by the type of direct_inf_model.

# Construct a Turing model
Z_t = randn(100)
latent_inf = generate_latent_infs(direct_inf_model, Z_t)

Now we can use the Turing PPL API to sample underlying parameters and generate the unobserved infections.

# Sample from the unobserved infections model

#Sample random parameters from prior
θ = rand(latent_inf)
#Get unobserved infections as a generated quantities from the model
I_t = generated_quantities(latent_inf, θ)

Fields

data::EpiData: Epidata object.

initialisation_prior::Distributions.Sampleable: Prior distribution for the initialisation of the infections. Default is Normal().

EpiAware.EpiData — Type

struct EpiData{T<:Real, F<:Function}

The EpiData struct represents epidemiological data used in infectious disease modeling.

Constructors

EpiData(gen_int, transformation::Function). Constructs an EpiData object with discrete

generation interval gen_int and transformation function transformation.

EpiData(;gen_distribution::ContinuousDistribution, D_gen, Δd = 1.0, transformation::Function = exp).

Constructs an EpiData object with double interval censoring discretisation of the continuous next generation interval distribution gen_distribution with additional right truncation at D_gen. Δd sets the interval width (default = 1.0). transformation sets the transformation function

Examples

Construction direct from discrete generation interval and transformation function:

using EpiAware
gen_int = [0.2, 0.3, 0.5]
g = exp
data = EpiData(gen_int, g)

Construction from continuous distribution for generation interval.

using Distributions

gen_distribution = Uniform(0.0, 10.0)

data = EpiData(;gen_distribution
    D_gen = 10.0)

Fields

gen_int::Vector{T} where T<:Real: Discrete generation interval.

len_gen_int::Integer: Length of the discrete generation interval.

transformation::Function: Transformation function defining constrained and unconstrained domain bijections.

EpiAware.ExpGrowthRate — Type

struct ExpGrowthRate{S<:Distributions.Sampleable} <: AbstractEpiModel

Model unobserved/latent infections as due to time-varying exponential growth rate $r_t$ which is generated by a latent process.

Mathematical specification

If $Z_t$ is a realisation of the latent model, then the unobserved/latent infections are given by

\[I_t = g(\hat{I}_0) \exp(Z_t).\]

where $g$ is a transformation function and the unconstrained initial infections $\hat{I}_0$ are sampled from a prior distribution, initialisation_prior which must be supplied to the DirectInfections constructor. The default initialisation_prior is the standard Normal Distributions.Normal().

Constructor

ExpGrowthRate can be constructed by passing an EpiData object and and subtype of Distributions.Sampleable.

Example usage with generate_latent_infs

generate_latent_infs can be used to construct a Turing model for the latent infections conditional on the sample path of a latent process. In this example, we generate a sample of a white noise latent process.

First, we construct an ExpGrowthRate struct with an EpiData object, an initialisation prior and a transformation function.

using Distributions, Turing, EpiAware
gen_int = [0.2, 0.3, 0.5]
g = exp

# Create an EpiData object
data = EpiData(gen_int, g)

# Create an ExpGrowthRate model
exp_growth_model = ExpGrowthRate(data = data, initialisation_prior = Normal())

Then, we can use generate_latent_infs to construct a Turing model for the unobserved infection generation model set by the type of direct_inf_model.

# Construct a Turing model
Z_t = randn(100) * 0.05
latent_inf = generate_latent_infs(exp_growth_model, Z_t)

Now we can use the Turing PPL API to sample underlying parameters and generate the unobserved infections.

# Sample from the unobserved infections model

#Sample random parameters from prior
θ = rand(latent_inf)
#Get unobserved infections as a generated quantities from the model
I_t = generated_quantities(latent_inf, θ)

Fields

data::EpiData

initialisation_prior::Distributions.Sampleable

EpiAware.Renewal — Type

struct Renewal{S<:Distributions.Sampleable} <: AbstractEpiModel

Model unobserved/latent infections as due to time-varying Renewal model with reproduction number $\mathcal{R}_t$ which is generated by a latent process.

Mathematical specification

If $Z_t$ is a realisation of the latent model, then the unobserved/latent infections are given by

\[\begin{align} \mathcal{R}_t &= g(Z_t),\\ I_t &= \mathcal{R}_t \sum_{i=1}^{n-1} I_{t-i} g_i, \qquad t \geq 1, \\ I_t &= g(\hat{I}_0) \exp(r(\mathcal{R}_1) t), \qquad t \leq 0. \end{align}\]

where $g$ is a transformation function and the unconstrained initial infections $\hat{I}_0$ are sampled from a prior distribution, initialisation_prior which must be supplied to the DirectInfections constructor. The default initialisation_prior is the standard Normal Distributions.Normal(). The discrete generation interval is given by $g_i$.

$r(\mathcal{R}_1)$ is the exponential growth rate implied by $\mathcal{R}_1)$ using the implicit relationship between the exponential growth rate and the reproduction number.

\[\mathcal{R} \sum_{j \geq 1} g_j \exp(- r j)= 1.\]

Constructor

Renewal can be constructed by passing an EpiData object and and subtype of Distributions.Sampleable.

Example usage with generate_latent_infs

generate_latent_infs can be used to construct a Turing model for the latent infections conditional on the sample path of a latent process. In this example, we generate a sample of a white noise latent process.

First, we construct an Renewal struct with an EpiData object, an initialisation prior and a transformation function.

using Distributions, Turing, EpiAware
gen_int = [0.2, 0.3, 0.5]
g = exp

# Create an EpiData object
data = EpiData(gen_int, g)

# Create an Renewal model
renewal_model = Renewal(data = data, initialisation_prior = Normal())

Then, we can use generate_latent_infs to construct a Turing model for the unobserved infection generation model set by the type of direct_inf_model.

# Construct a Turing model
Z_t = randn(100) * 0.05
latent_inf = generate_latent_infs(renewal_model, Z_t)

Now we can use the Turing PPL API to sample underlying parameters and generate the unobserved infections.

# Sample from the unobserved infections model

#Sample random parameters from prior
θ = rand(latent_inf)
#Get unobserved infections as a generated quantities from the model
I_t = generated_quantities(latent_inf, θ)

Fields

data::EpiData

initialisation_prior::Distributions.Sampleable

EpiAware.Renewal — Method

Callable on a Renewal struct for compute new incidence based on recent incidence and Rt.

Mathematical specification

The new incidence is given by

\[I_t = R_t \sum_{i=1}^{n-1} I_{t-i} g_i\]

where I_t is the new incidence, R_t is the reproduction number, I_{t-i} is the recent incidence and g_i is the generation interval.

Arguments

recent_incidence: Array of recent incidence values.

Rt: Reproduction number.

Returns

Tuple containing the updated incidence array and the new incidence value.

EpiAware.R_to_r — Method

R_to_r(
    R₀,
    w::Array{T<:AbstractFloat, 1};
    newton_steps,
    Δd
) -> Any

This function computes an approximation to the exponential growth rate r given the reproductive ratio R₀ and the discretized generation interval w with discretized interval width Δd. This is based on the implicit solution of

\[G(r) - {1 \over R_0} = 0.\]

where

\[G(r) = \sum_{i=1}^n w_i e^{-r i}.\]

is the negative moment generating function (MGF) of the generation interval distribution.

The two step approximation is based on: 1. Direct solution of implicit equation for a small r approximation. 2. Improving the approximation using Newton's method for a fixed number of steps newton_steps.

Returns:

The approximate value of r.

EpiAware.create_discrete_pmf — Method

create_discrete_pmf(
    dist::Distributions.Distribution,
    ::Val{:single_censored};
    primary_approximation_point,
    Δd,
    D
)

Create a discrete probability mass function (PMF) from a given distribution, assuming that the primary event happens at primary_approximation_point * Δd within an intial censoring interval. Common single-censoring approximations are primary_approximation_point = 0 (left-hand approximation), primary_approximation_point = 1 (right-hand) and primary_approximation_point = 0.5 (midpoint).

Arguments:

dist: The distribution from which to create the PMF.

::Val{:single_censored}: A dummy argument to dispatch to this method. The purpose of the Val

type argument is that to use single-censored approximation is an active decision.

primary_approximation_point: A approximation point for the primary time in its censoring interval.

Default is 0.5 for midpoint approximation.

Δd: The step size for discretizing the domain. Default is 1.0.

D: The upper bound of the domain. Must be greater than Δd.

Returns:

A vector representing the PMF.

Raises:

AssertionError if the minimum value of dist is negative.

AssertionError if Δd is not positive.

AssertionError if D is not greater than Δd.

EpiAware.create_discrete_pmf — Method

create_discrete_pmf(dist::Distributions.Distribution; Δd, D)

Create a discrete probability mass function (PMF) from a given distribution, assuming a uniform distribution over primary event times with censoring intervals of width Δd for both primary and secondary events. The CDF for the time from the left edge of the interval containing the primary event to the secondary event is created by direct numerical integration of the convolution of the CDF of dist with the uniform density on [0,Δd), the discrete PMF for double censored delays is then found using simple differencing on the CDF.

Arguments:

dist: The distribution from which to create the PMF.

Δd: The step size for discretizing the domain. Default is 1.0.

D: The upper bound of the domain. Must be greater than Δd.

Returns:

A vector representing the PMF.

Raises:

AssertionError if the minimum value of dist is negative.

AssertionError if Δd is not positive.

AssertionError if D is not greater than Δd.

EpiAware.generate_latent_infs — Method

generate_latent_infs(
    epi_model::AbstractEpiModel,
    Z_t
) -> Any

Generate unobserved/latent infections based on the given epi_model <: AbstractEpimodel and a latent process path $Z_t$.

The generate_latent_infs function implements a model of generating unobserved/latent infections conditional on a latent process. Which model of generating unobserved/latent infections to be implemented is set by the type of epi_model. If no implemention is defined for the given epi_model, then EpiAware will return a warning and return nothing.

Interface to Turing.jl probablilistic programming language (PPL)

Apart from the no implementation fallback method, the generate_latent_infs implementation function should be a constructor function for a DynamicPPL.Model object.

EpiAware.generate_latent_infs — Method

generate_latent_infs(
    epi_model::DirectInfections,
    Z_t
) -> Any

Implement the generate_latent_infs function for the DirectInfections model.

Example usage with DirectInfections type of model for unobserved infection process

First, we construct a DirectInfections struct with an EpiData object, an initialisation prior and a transformation function.

using Distributions, Turing, EpiAware
gen_int = [0.2, 0.3, 0.5]
g = exp

# Create an EpiData object
data = EpiData(gen_int, g)

# Create a DirectInfections model
direct_inf_model = DirectInfections(data = data, initialisation_prior = Normal())

Then, we can use generate_latent_infs to construct a Turing model for the unobserved infection generation model set by the type of direct_inf_model.

# Construct a Turing model
Z_t = randn(100)
latent_inf = generate_latent_infs(direct_inf_model, Z_t)

Now we can use the Turing PPL API to sample underlying parameters and generate the unobserved infections.

# Sample from the unobserved infections model

#Sample random parameters from prior
θ = rand(latent_inf)
#Get unobserved infections as a generated quantities from the model
I_t = generated_quantities(latent_inf, θ)

EpiAware.generate_latent_infs — Method

generate_latent_infs(epi_model::ExpGrowthRate, rt) -> Any

Implement the generate_latent_infs function for the ExpGrowthRate model.

Example usage with ExpGrowthRate type of model for unobserved infection process

generate_latent_infs can be used to construct a Turing model for the latent infections conditional on the sample path of a latent process. In this example, we generate a sample of a white noise latent process.

First, we construct an ExpGrowthRate struct with an EpiData object, an initialisation prior and a transformation function.

using Distributions, Turing, EpiAware
gen_int = [0.2, 0.3, 0.5]
g = exp

# Create an EpiData object
data = EpiData(gen_int, g)

# Create an ExpGrowthRate model
exp_growth_model = ExpGrowthRate(data = data, initialisation_prior = Normal())

Then, we can use generate_latent_infs to construct a Turing model for the unobserved infection generation model set by the type of direct_inf_model.

# Construct a Turing model
Z_t = randn(100) * 0.05
latent_inf = generate_latent_infs(exp_growth_model, Z_t)

Now we can use the Turing PPL API to sample underlying parameters and generate the unobserved infections.

# Sample from the unobserved infections model

#Sample random parameters from prior
θ = rand(latent_inf)
#Get unobserved infections as a generated quantities from the model
I_t = generated_quantities(latent_inf, θ)

EpiAware.generate_latent_infs — Method

generate_latent_infs(epi_model::Renewal, _Rt) -> Any

Implement the generate_latent_infs function for the Renewal model.

Example usage with Renewal type of model for unobserved infection process

generate_latent_infs can be used to construct a Turing model for the latent infections conditional on the sample path of a latent process. In this example, we generate a sample of a white noise latent process.

First, we construct an Renewal struct with an EpiData object, an initialisation prior and a transformation function.

using Distributions, Turing, EpiAware
gen_int = [0.2, 0.3, 0.5]
g = exp

# Create an EpiData object
data = EpiData(gen_int, g)

# Create an Renewal model
renewal_model = Renewal(data = data, initialisation_prior = Normal())

Then, we can use generate_latent_infs to construct a Turing model for the unobserved infection generation model set by the type of renewal_model.

# Construct a Turing model
Z_t = randn(100) * 0.05
latent_inf = generate_latent_infs(renewal_model, Z_t)

Now we can use the Turing PPL API to sample underlying parameters and generate the unobserved infections.

# Sample from the unobserved infections model

#Sample random parameters from prior
θ = rand(latent_inf)
#Get unobserved infections as a generated quantities from the model
I_t = generated_quantities(latent_inf, θ)

EpiAware.manypathfinder — Method

manypathfinder(
    mdl::DynamicPPL.Model,
    ndraws;
    nruns,
    maxiters,
    max_tries,
    kwargs...
) -> Any

Run multiple instances of the pathfinder algorithm and returns the pathfinder run with the largest ELBO estimate.

Arguments

mdl::DynamicPPL.Model: The model to perform inference on.

nruns::Int: The number of pathfinder runs to perform.

ndraws::Int: The number of draws per pathfinder run, readjusted to be at least as large as the number of chains.

nchains::Int: The number of chains that will be initialised by pathfinder draws.

maxiters::Int: The maximum number of optimizer iterations per pathfinder run.

max_tries::Int: The maximum number of extra tries to find a valid pathfinder result.

kwargs...: Additional keyword arguments passed to pathfinder.

Returns

best_pfs::PathfinderResult: Best pathfinder result by estimated ELBO.

EpiAware.r_to_R — Method

r_to_R(r, w::AbstractVector) -> Any

r_to_R(r, w)

Compute the reproductive ratio given exponential growth rate r and discretized generation interval w.

Arguments

r: The exponential growth rate.

w: discretized generation interval.

Returns

The reproductive ratio.

EpiAware.scan — Method

scan(f::AbstractModel, init, xs) -> Tuple{Any, Any}

Apply f to each element of xs and accumulate the results.

f must be a callable on a sub-type of AbstractModel.

Design note

scan is being restricted to AbstractModel sub-types to ensure: 1. That compiler specialization is activated 2. Also avoids potential compiler overhead from specialisation on f<: Function.

Arguments

f: A callable/functor that takes two arguments, carry and x, and returns a new carry and a result y.

init: The initial value for the carry variable.

xs: An iterable collection of elements.

Returns

ys: An array containing the results of applying f to each element of xs.

carry: The final value of the carry variable after processing all elements of xs.

EpiAware.spread_draws — Method

spread_draws(chn::MCMCChains.Chains) -> DataFrames.DataFrame

spread_draws(chn::Chains)

Converts a Chains object into a DataFrame in tidybayes format.

Arguments

chn::Chains: The Chains object to be converted.

Returns

df::DataFrame: The converted DataFrame.