Public Documentation

Documentation for EpiAwareBae.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

Contents

Contents

Index

Public API

Index

EpiAware.EpiAwareUtils

EpiAware.EpiAwareUtils.DirectSample

EpiAware.EpiAwareUtils.HalfNormal

EpiAware.EpiAwareUtils.censored_pmf

EpiAware.EpiAwareUtils.censored_pmf

EpiAware.EpiAwareUtils.get_param_array

EpiAware.EpiAwareUtils.prefix_submodel

EpiAware.EpiAwareUtils.scan

EpiAware.EpiAwareUtils.spread_draws

Public API

EpiAware.EpiAwareUtils — Module

Module for defining utility functions.

EpiAware.EpiAwareUtils.DirectSample — Type

struct DirectSample <: AbstractEpiSamplingMethod

Sample directly from a Turing model.

Fields

n_samples::Union{Nothing, Int64}: Number of samples from a model. If an integer is provided, the model is sampled n_samples times using Turing.Prior() returning an MCMChains. Chain object. If nothing, the model is sampled once returning a NamedTuple object of the sampled random variables along with generated quantities

EpiAware.EpiAwareUtils.HalfNormal — Type

struct HalfNormal{T<:Real} <: Distributions.Distribution{Distributions.Univariate, Distributions.Continuous}

Create a half-normal prior distribution with the specified mean.

Arguments:

μ: The mean of the half-normal distribution.

Returns:

A HalfNormal distribution with the specified mean.

Examples:

using EpiAware, Distributions

hn = HalfNormal(1.0)
# output
EpiAware.EpiAwareUtils.HalfNormal{Float64}(μ=1.0)

filter out all the values that are less than 0

rand(hn)
# output
0.4508533245229199

cdf(hn, 2)
# output
0.8894596502772643

quantile(hn, 0.5)
# output
0.8453475393951495

logpdf(hn, 2)
# output
-3.1111166111445083

mean(hn)
# output
1.0

var(hn)
# output
0.5707963267948966

Fields

μ::Real

EpiAware.EpiAwareUtils.censored_pmf — Method

censored_pmf(
    dist::Distributions.Distribution,
    ::Val{:single_censored};
    primary_approximation_point,
    Δd,
    D
)

Create a discrete probability mass function (PMF) from a given distribution, assuming that the primary event happens at primary_approximation_point * Δd within an intial censoring interval. Common single-censoring approximations are primary_approximation_point = 0 (left-hand approximation), primary_approximation_point = 1 (right-hand) and primary_approximation_point = 0.5 (midpoint).

Arguments

dist: The distribution from which to create the PMF.

::Val{:single_censored}: A dummy argument to dispatch to this method. The purpose of the Val

type argument is that to use single-censored approximation is an active decision.

primary_approximation_point: A approximation point for the primary time in its censoring interval.

Default is 0.5 for midpoint approximation.

Δd: The step size for discretizing the domain. Default is 1.0.

D: The upper bound of the domain. Must be greater than Δd.

Returns

A vector representing the PMF.

Raises:

AssertionError if the minimum value of dist is negative.

AssertionError if Δd is not positive.

AssertionError if D is not greater than Δd.

Examples

using Distributions
using EpiAware.EpiAwareUtils

dist = Exponential(1.0)

censored_pmf(dist, Val(:single_censored); D = 10) |>
    p -> round.(p, digits=3)

# output
10-element Vector{Float64}:
 0.393
 0.383
 0.141
 0.052
 0.019
 0.007
 0.003
 0.001
 0.0
 0.0

EpiAware.EpiAwareUtils.censored_pmf — Method

censored_pmf(dist::Distributions.Distribution; Δd, D) -> Any

Create a discrete probability mass function (PMF) from a given distribution, assuming a uniform distribution over primary event times with censoring intervals of width Δd for both primary and secondary events. The CDF for the time from the left edge of the interval containing the primary event to the secondary event is created by direct numerical integration of the convolution of the CDF of dist with the uniform density on [0,Δd), the discrete PMF for double censored delays is then found using simple differencing on the CDF.

Arguments

dist: The distribution from which to create the PMF.

Δd: The step size for discretizing the domain. Default is 1.0.

D: The upper bound of the domain. Must be greater than Δd. Default D = nothing

indicates that the distribution should be truncated at its 99th percentile rounded to nearest multiple of Δd.

Returns

A vector representing the PMF.

Raises

AssertionError if the minimum value of dist is negative.

AssertionError if Δd is not positive.

AssertionError if D is shorter than Δd.

AssertionError if D is not a multiple of Δd.

Examples

using Distributions
using EpiAware.EpiAwareUtils

dist = Exponential(1.0)

censored_pmf(dist; D = 10) |>
    p -> round.(p, digits=3)

# output
10-element Vector{Float64}:
 0.368
 0.4
 0.147
 0.054
 0.02
 0.007
 0.003
 0.001
 0.0
 0.0

EpiAware.EpiAwareUtils.get_param_array — Method

get_param_array(chn::MCMCChains.Chains) -> Any

Extract a parameter array from a Chains object chn that matches the shape of number of sample and chain pairs in chn.

Arguments

chn::Chains: The Chains object containing the MCMC samples.

Returns

param_array: An array of parameter samples, where each element corresponds to a single

MCMC sample as a NamedTuple.

Example

Sampling from a simple model which has both scalar and vector quantity random variables across 4 chains.

using Turing, MCMCChains, EpiAware

@model function testmodel()
    y ~ Normal()
end
mdl = testmodel()
chn = sample(mdl, Prior(), MCMCSerial(), 2, 1, progress=false)

A = get_param_array(chn)

EpiAware.EpiAwareUtils.prefix_submodel — Method

prefix_submodel(
    model::AbstractModel,
    fn::Function,
    prefix::String,
    kwargs...
) -> Any

Generate a submodel with an optional prefix. A lightweight wrapper around the @submodel macro from DynamicPPL.jl.

Arguments

model::AbstractModel: The model to be used.

fn::Function: The Turing @model function to be applied to the model.

prefix::String: The prefix to be used. If the prefix is an empty string, the submodel is created without a prefix.

Returns

submodel: The returns from the submodel are passed through.

Examples

using EpiAware, DynamicPPL

submodel = prefix_submodel(FixedIntercept(0.1), generate_latent, string(1), 2)
submodel
# output
Model{typeof(prefix_submodel), (:model, :fn, :prefix, Symbol("#splat#kwargs")), (), (), Tuple{FixedIntercept{Float64}, typeof(generate_latent), String, Tuple{Int64}}, Tuple{}, DefaultContext}(EpiAware.EpiAwareUtils.prefix_submodel, (model = FixedIntercept{Float64}(0.1), fn = EpiAware.EpiAwareBase.generate_latent, prefix = "1", var"#splat#kwargs" = (2,)), NamedTuple(), DefaultContext())

We can now draw a sample from the submodel.

rand(submodel)

EpiAware.EpiAwareUtils.scan — Method

scan(f::AbstractModel, init, xs) -> Tuple{Any, Any}

Apply f to each element of xs and accumulate the results.

f must be a callable on a sub-type of AbstractModel.

Design note

scan is being restricted to AbstractModel sub-types to ensure: 1. That compiler specialization is activated 2. Also avoids potential compiler overhead from specialisation on f<: Function.

Arguments

f: A callable/functor that takes two arguments, carry and x, and returns a new carry and a result y.

init: The initial value for the carry variable.

xs: An iterable collection of elements.

Returns

ys: An array containing the results of applying f to each element of xs.

carry: The final value of the carry variable after processing all elements of xs.

Examples

```jldoctest using EpiAware

struct Adder <: EpiAwareBase.AbstractModel end function (a::Adder)(carry, x) carry + x, carry + x end

scan(Adder(), 0, 1:5) #output ([1, 3, 6, 10, 15], 15)

EpiAware.EpiAwareUtils.spread_draws — Method

spread_draws(chn::MCMCChains.Chains) -> DataFrames.DataFrame

spread_draws(chn::Chains)

Converts a Chains object into a DataFrame in tidybayes format.

Arguments

chn::Chains: The Chains object to be converted.

Returns

df::DataFrame: The converted DataFrame.