internal Documentation

Documentation for EpiAware.jl's internal interface.

Contents

Index

EpiAware.NegativeBinomialMeanClust

EpiAware._continue_manypathfinder!

EpiAware._get_best_elbo_pathfinder

EpiAware._run_manypathfinder

EpiAware.generate_observation_kernel

EpiAware.neg_MGF

Internal API

EpiAware.NegativeBinomialMeanClust — Method

NegativeBinomialMeanClust(
    μ,
    α
) -> Distributions.NegativeBinomial

NegativeBinomialMeanClust(μ, α)

Compute the mean-cluster factor negative binomial distribution.

Arguments

μ: The mean of the distribution.

α: The clustering factor parameter.

Returns

A NegativeBinomial distribution object.

Methods

NegativeBinomialMeanClust(μ, α)

defined at /home/runner/work/Rt-without-renewal/Rt-without-renewal/src/utils/dist.jl:164.

EpiAware._continue_manypathfinder! — Method

_continue_manypathfinder!(
    pfs,
    mdl::DynamicPPL.Model;
    max_tries,
    nruns,
    kwargs...
)

Continue running the pathfinder algorithm until a pathfinder succeeds or the maximum number of tries is reached.

Arguments

pfs: An array of pathfinder objects.

mdl::DynamicPPL.Model: The model to perform inference on.

max_tries: The maximum number of tries to run the pathfinder algorithm. Default is Inf.

nruns: The number of times to run the pathfinder function.

kwargs...: Additional keyword arguments passed to pathfinder.

Returns

pfs: The updated array of pathfinder objects.

Methods

_continue_manypathfinder!(
    pfs,
    mdl;
    max_tries,
    nruns,
    kwargs...
)

defined at /home/runner/work/Rt-without-renewal/Rt-without-renewal/src/inferencemethods/manypathfinder.jl:41.

EpiAware._get_best_elbo_pathfinder — Method

_get_best_elbo_pathfinder(pfs) -> Any

Selects the pathfinder with the highest ELBO estimate from a list of pathfinders.

Arguments

pfs: A list of pathfinders results or Symbol values indicating failure.

Returns

The pathfinder with the highest ELBO estimate.

Methods

_get_best_elbo_pathfinder(pfs)

defined at /home/runner/work/Rt-without-renewal/Rt-without-renewal/src/inferencemethods/manypathfinder.jl:71.

EpiAware._run_manypathfinder — Method

_run_manypathfinder(mdl::DynamicPPL.Model; nruns, kwargs...)

Run pathfinder multiple times and store the results in an array. Fails safely.

Arguments

mdl::DynamicPPL.Model: The Turing model to be used for inference.

nruns: The number of times to run the pathfinder function.

kwargs...: Additional keyword arguments passed to pathfinder.

Returns

An array of PathfinderResult objects or Symbol values indicating success or failure.

Methods

_run_manypathfinder(mdl; nruns, kwargs...)

defined at /home/runner/work/Rt-without-renewal/Rt-without-renewal/src/inferencemethods/manypathfinder.jl:12.

EpiAware.generate_observation_kernel — Method

generate_observation_kernel(delay_int, time_horizon) -> Any

Generate an observation kernel matrix based on the given delay interval and time horizon.

Arguments

delay_int::Vector{Float64}: The delay PMF vector.

time_horizon::Int: The number of time steps of the observation period.

Returns

K::SparseMatrixCSC{Float64, Int}: The observation kernel matrix.

Methods

generate_observation_kernel(delay_int, time_horizon)

defined at /home/runner/work/Rt-without-renewal/Rt-without-renewal/src/observationmodels/delayobservations.jl:11.

EpiAware.neg_MGF — Method

neg_MGF(r, w::AbstractVector) -> Any

Compute the negative moment generating function (MGF) for a given rate r and weights w.

Arguments

r: The rate parameter.

w: An abstract vector of weights.

Returns

The value of the negative MGF.

Methods

neg_MGF(r, w)

defined at /home/runner/work/Rt-without-renewal/Rt-without-renewal/src/utils/dist.jl:88.