internal Documentation

Documentation for EpiAware.jl's internal interface.

Contents

Index

EpiAware.R_to_r

EpiAware.generate_latent_infs

EpiAware.generate_observation_kernel

EpiAware.mean_cc_neg_bin

EpiAware.neg_MGF

EpiAware.r_to_R

Internal API

EpiAware.R_to_r — Method

R_to_r(R₀, w::Vector{T}; newton_steps = 2, Δd = 1.0)

This function computes an approximation to the exponential growth rate r given the reproductive ratio R₀ and the discretized generation interval w with discretized interval width Δd. This is based on the implicit solution of

\[G(r) - {1 \over R_0} = 0.\]

where

\[G(r) = \sum_{i=1}^n w_i e^{-r i}.\]

is the negative moment generating function (MGF) of the generation interval distribution.

The two step approximation is based on: 1. Direct solution of implicit equation for a small r approximation. 2. Improving the approximation using Newton's method for a fixed number of steps newton_steps.

Returns:

The approximate value of r.

EpiAware.generate_latent_infs — Method

generate_latent_infs(epi_model::Renewal, _Rt)

Turing model constructor for latent infections using the Renewal object epi_model and time-varying unconstrained reproduction number _Rt.

generate_latent_infs creates a Turing model for sampling latent infections with given unconstrainted reproduction number _Rt but random initial incidence scale. The initial incidence pre-time one is given as a scale on top of an exponential growing process with exponential growth rate given by R_to_rapplied to the first value of Rt.

Arguments

epi_model::Renewal: The epidemiological model.

_Rt: Time-varying unconstrained (e.g. log-) reproduction number.

Returns

I_t: Array of latent infections over time.

EpiAware.generate_observation_kernel — Method

generate_observation_kernel(delay_int, time_horizon)

Generate an observation kernel matrix based on the given delay interval and time horizon.

Arguments

delay_int::Vector{Float64}: The delay PMF vector.

time_horizon::Int: The number of time steps of the observation period.

Returns

K::SparseMatrixCSC{Float64, Int}: The observation kernel matrix.

EpiAware.mean_cc_neg_bin — Method

mean_cc_neg_bin(μ, α)

Compute the mean-cluster factor negative binomial distribution.

Arguments

μ: The mean of the distribution.

α: The clustering factor parameter.

Returns

A NegativeBinomial distribution object.

EpiAware.neg_MGF — Method

neg_MGF(r, w::AbstractVector)

Compute the negative moment generating function (MGF) for a given rate r and weights w.

Arguments

r: The rate parameter.

w: An abstract vector of weights.

Returns

The value of the negative MGF.

EpiAware.r_to_R — Method

r_to_R(r, w)

Compute the reproductive ratio given exponential growth rate r and discretized generation interval w.

Arguments

r: The exponential growth rate.

w: discretized generation interval.

Returns

The reproductive ratio.