Public Documentation

Documentation for EpiAware.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

Contents

Contents

Index

Public API

Index

EpiAware.EpiAware

Public API

EpiAware.EpiAware — Module

module EpiAware

EpiAware provides functionality for fitting epidemiological models to data. It is built on top of the Turing probabilistic programming language, and provides a set of utilities for constructing and fitting models to data.

Core model structure

An epidemiological model in EpiAware consists of composable structs with core abstract types. The core types are:

AbstractModel: This overarching type is used to abstract Turing models and is inherited by the other abstract types we use.

AbstractTuringEpiModel: Subtypes of this abstract type represent different models for the spread of an infectious disease. Each model type has a corresponding make_epi_aware function that constructs a Turing model for fitting the model to data. Implemented concrete subtypes:

Renewal: A renewal process model for the spread of an infectious disease.

ExpGrowthRate: An exponential growth rate model for the spread of an infectious disease.

DirectInfections: A model for the spread of an infectious disease based on modelling direct infections.

AbstractLatentModel: Subtypes of this abstract type represent different latent processes that can be used in an epidemiological model. Implemented concrete subtype:

RandomWalk: A random walk latent process.

AbstractObservationModel: Subtypes of this abstract type represent different observation models that can be used in an epidemiological model. Implemented concrete subtypes:

DelayObservation: An observation process that models the delay between the time of infection and the time of observation as a convolution, followed by a negative binomial distributed sample.